handler api

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,5 @@
 using System;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -6,18 +7,26 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivity
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public string Id { get; set; }
     }
 
-    public class Handler(AppDbContext context) : IRequestHandler<Command>
+    public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
     {
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
             var activity = await context.Activities.FindAsync([request.Id], cancellationToken) ?? throw new Exception("Atividade não encontrada");
+
+            if (activity == null) return Result<Unit>.Failure("Activity not fount", 404);
+
             context.Activities.Remove(activity);
-            await context.SaveChangesAsync(cancellationToken);
+
+            var result = await context.SaveChangesAsync(cancellationToken) > 0;
+
+            if (!result) return Result<Unit>.Failure("Failed to delete the activity", 400);
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
